crostini: Opt in to playstore in precondition
Starting up ARC++ with a real gaia login requires opting into the
playstore. This means the RunWithArc test will fail for not opting in,
but it's attempt to boot Android will (through unknown mechanisms)
prevent ShareFolders from being able to opt-in.

Additionally, trying to opt-in twice fails because the opt-in page
doesn't appear the second time. This means we can't just make
RunWithArc run the opt-in flow. Instead we must move the opt-in to the
precondition. Tests should just use arc.New().

BUG=chromium:1128743, chromiut:1154444
TEST=RunWithArc, ShareMovies

Disallow-Recycled-Builds: test-failures
Change-Id: I17679cdbb04f4dd690b446e3753664a6c157df10