crostini: Register LXD server component even when using DLC
To force the test VM to use a preloaded test container image,
instead of the production image, we register the
cros-crostini-image-server-url component. This was not being done when
the install mode was set to DLC, causing DLC tests to use the prod
container image. This causes several tests which require certain
pre-installed packages in the container to fail.

BUG=chromium:1128743
TEST=AudioBasic, AudioPlaybackConfigurations, CopyPaste, DragDrop,
GPUEnabled, SecureCopyPaste, VmcExtraDisk, Webserver

Disallow-Recycled-Builds: test-failures
Change-Id: I83fc2879a3d555c60c8b9d4741270e862ab6adc1