tast-tests: Wait for clamshell split view resize update after mouse move
The present CL changes how ui.SplitViewResizePerf.clamshell_mode deals
with the fact that a window resize can take a moment to update after the
mouse moves. If the mouse button is released in that moment, then the
resize will not be updated for the final position of the mouse.

BEFORE: There is code that checks where the resize actually ended, in
preparation for a subsequent resize to begin from there. In case of
terrible clamshell split view resizing performance, the test can fail
because the resize ends so far right that the window becomes maximized.

AFTER: Wait for the resize to finish before releasing the mouse button.
There is no need for code that checks where the resize actually ended.

BUG=chromium:1162184,chromium:1164562
TEST=ui.SplitViewResizePerf.clamshell_mode

Change-Id: Ied434174a83c7e95c00115f2f4407a6c12cf0f04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2625985
Tested-by: Avery Musbach <amusbach@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>