firmware: Allow bio_fw_updater previous log to not exist
In actual lab tests it's often the case that the test runs at the
first boot after installing an OS image, and fingerprint firmware
update is not needed, therefore there's no bio_fw_updater.PREVIOUS
log file. The test has been failing very often due to this. Handle
this case in the test.

BUG=b:169958710
TEST=remove bio_fw_updater.PREVIOUS on DUT and
     tast -verbose run <IP> firmware.FpUpdaterSucceeded

Change-Id: Ic176b13593deaba97d04a1794d1345d5636c5066