crostini: Disable usage of the real gaia account
In CL:2569009 we switched to using a real gaia account by default,
which appears to have introduced flake in setting up chrome.

Disable this for now, since no cq tests actually need it. This will
break ShareDrive, ShareMovies, and NoAccessToDrive, but none of those
tests are on the CQ currently.

BUG=b:177396586, chromium:1166530
TEST=CQ

Disallow-Recycled-Builds: test-failures
Change-Id: I13a5247233ecf6a8b37067d47c16833f127047f9