ui: FocusAndWait change to a global wait
In the Crostini tests we are hitting an error on a few boards where the
focus is performed, but the focus event is never received.

I suspect this is because the accessibility node is destroyed and
recreated between the focus operation and the focus event being emitted.

Change to wait for any focus event in the system, rather than the
particular node that may have been destroyed before the focus event was
sent.

BUG=None
TEST=Ran crostini.Basic which exercises the focusing code

Change-Id: Iae5c3e5be72a8a128b3d0c4a58ca30c9b2df5b8f