chrome: Move Conn and JSObject to an internal package
This patch introduces a new internal package "driver", which focuses on
interacting with a running Chrome process. This package does not handle
setting up / starting Chrome processes.

As a first step, move Conn and JSObject to the driver package.

This is a refactoring change. No behavior change is expected.

BUG=chromium:1166043
TEST=fast_build.sh -T
TEST=tast run DUT ui.ChromeLogin

Change-Id: I39d16caf79c39f63087e62a6618e42fd04bedaaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2632550
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>