tast-test: Reverse patchpanel addresses
The local proxy server is running in an isolated network namespace,
created by patchpanel. Patchpanel allocates the host and peer addresses
on the virtual network interfaces associated with the proxy.

CL:2568913 fixed a bug where the host and peer address were reversed.
This CL updates the tast proxy server to reflect this change.

BUG=b:177840593
TEST=tast run $DUT network.TestProxyServer

Change-Id: I7b05997f4ae96315251f6183d8a516c07a92bc92