tast-tests: Add apps.LaunchSystemWebAppsFromURL
Increasing the test coverage for System Web Apps (SWAs) we try to launch
all the SWAs by typing their chrome:// URL into the Chrome omnibox. This
does not use |conn.Navigate| or |cr.NewConn| to navigate as it utilises
DevTools to perform the navigation which does not verify link capturing
in SWAs.

BUG=chromium:1146904
TEST=tast run <DUT-IP> apps.LaunchSystemWebAppsFromURL

Change-Id: I48b23bbc7408066ccf55a0b5588e1647acc3ccb4