tast-tests: stabilize ui.InputMethodManagement
ui.InputMethodManagement is flaky sometimes, throwing error:
`Failed to launch OS settings and land at inputs setting page: context deadline exceeded; last error follows: failed to match condition after changing page location in javascript`

In the screenshot, it is still on the main settings page, which means the javascript function to change window.location is not executed successfully, and no error is thrown out.

Solution: Since this is support function working around UI to navigate to the feature setting page, which is not part of the test journey. The proposed improvement is to add fail retry on condition check.

BUG=b:177868367
TEST=ui.InputMethodManagement

Change-Id: I78ee8a47c06b0670e5c99306e72fba128c948608