crostini: prevent panic on failed uninstall
In the uninstallation part of precondition we use log instead of fatal
so that cleanup failure will not fail the test.  However if there is
actually an error with one of the steps the log statements fall through
and we get a nil pointer dereference on the next step.

Move the uninstallation code to its own function and return after each
step failure so that we do not get any panics.

BUG=None
TEST=ran crostini.Basic on hatch

Change-Id: Iadfb5bb0026d8811811331e98621145a5976963f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2638958
Tested-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>