tast-tests: add base of uiauto library
This adds the base of the uiauto library. It adds the nodewith package to
generate a way to find nodes. It adds packages for each type of
constant. It still is missing many features to make it have parity with
the old ui library. ui.ChangeWallpaper is ported to show what the new
api looks like in use.

BUG=b:178020071
TEST=tast run <eve> ui.ChangeWallpaper

Change-Id: Ia6763ba4541458010e5a500a6433ae4752011811