scanner: Wait for CUPS autoconfiguration
If the user is logged in, CUPS will attempt to auto-configure newly
added IPP-USB devices as printers.  It will start ippusb_manager and
ippusb_bridge in the process, which then intereferes with the way tests
want those set up.

If tests win the race, they can block CUPS, but this produces occasional
flakes.  Instead, wait for CUPS to finish configuring the printer and
clean up the processes it started before running the scanning tests.

We now also have to ensure that Chrome is logged in before starting the
test because otherwise printer autoconfiguration won't be triggered.
This avoids another source of inconsistent behavior.

BUG=b:162003158
TEST=1000 iterations of tast run scanner.* on betty and zork

Change-Id: If7f0661f13f63438f4f5e5737c5a6695fcf830ca