Add support to install split APKs
This CL adds support to install split APKs.
"Split APKs" are generated when an App Bundle is published in Play
Store.
Whenever we need to test an 3rd party App Bundle, the installable is
going to be an "split APK".
In particular this CL is needed to test Roblox.

BUG=b:169182810
TEST=Tested that it doesn't break existing functionality
     adb install-multiple tested in follow-up CL

Change-Id: Iec114e43e6e46af543ab4ab01df04558f7840522
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2641814
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Auto-Submit: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>