crostini: Don't delete the DLC image
In CL:2639196 we started trying to cleanup the termina dlc when
closing the precondition to ensure other tests didn't rely on it by
accident. However, the function we called actually deletes the DLC
image off disk entirely, so once the precondition had been closed once
no other tests would be able to install crostini with DLC.

This is clearly undesirable, so use the DLC uninstall method
instead. Despite the name, this leaves the image on disk (at least,
for long enough for our purposes) and just removes the mount.

BUG=chromium:1128743
TEST=RemoveOk + ResizeCancel

Disallow-Recycled-Builds: test-failures
Change-Id: Ib3f70733df1549f704803ee39b92bee55482da8b