arc: Fix flaky PhysicalKeyboard test on betty (arc-r)
This adds wait for CPU is idle as a speculative fix for UIAutomator
based tests that may be affected by many heavy flows in background.

BUG=b:168184423
TEST=tast run  DUT arc.PhysicalKeyboard.vm. Note, ARC is not started on
betty on my Cloud instance so I could not confirm results on betty. If
this helps to stabilize the test, similar fix could be applied to other
UI based critical tests

Change-Id: Ia45aa7aca846dabf3f2a243e387b6ce9a2c9c712