Tast: Pass WifiCheckIntelSARTable on devices without SAR tables
The purpose of this test is to ensure that devices do not have
poorly formatted or otherwise faulty SAR tables. By definition, a
device without any SAR tables meets this criteria, so this change
makes the test automatically pass when no SAR tables are found.

BUG=b:175816675
TEST=`tast run ${DUT} network.WifiCheckIntelSARTable`
Test passes on devices without SAR tables.
Behaves as normal for devices with SAR tables.

Change-Id: I4f1edc641e0106a8cd3daf7536c1a155db4e80aa