tast-tests: Add apps.Sharesheet test
Sharesheet allows sharing not only to Android apps but also to
Progressive Web Apps (PWAs). This test installs a PWA that has a
registered web share target which should allow the Sharesheet to share a
file to the app. It also verifies the data was received correctly.

BUG=chromium:1169916
TEST=tast run <DUT-IP> apps.Sharesheet

Change-Id: Ic2811fc8ce455d2ef65d5413c72dae9bd4ce4896
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2643688
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>