video: factor caps detection logic out of video.Capability
This CL refactors video.Capability to take the caps detection
logic out of the test, so that in the following CL the camera
specific parts can be factored out.

No new functionality is intended, although I took a pass at
the comments to make them more understandable in light of
feedback gathered during bringups.

BUG=b:177261819
TEST=tast run video.Capability on Samus.

Change-Id: Ie45237690259110908c854c2dc8e6209d9f4fbf6