chrome: Save test extensions to persistent directory
Formerly Chrome.Close used to delete temporary directories where test
extensions are saved. This may cause Chrome to malfunction on post-test
manual inspection, as well as on reconnecting to Chrome, which is to be
implemented soon.

This patch changes the library to save extension files to the
"persistent" directory, which is not deleted on Chrome.Close. It is
instead cleared on chrome.New. Note that unpacked extensions specified
with chrome.UnpackedExtension are also copied to this directory.

BUG=chromium:1170186
TEST=fast_build.sh -T
TEST=tast run DUT example.ChromeExtension

Change-Id: Ic576967afa63c1b5bde8bbdc3d0b831dfabee1ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2644592
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>