webrtc: Set display to landscape orientation in peerconnection test
The default resolution produced by camera in RTC peerconnection
page depends on the display resolution. On scarlet, it is
1920x1080 and 1080x1920 if the display is in landscape and
portrait orientation, respectively. However, the maximum
resolution supported by the hardware encoder on scarlet is
1920x1088. Therefore, the hardware encoder is not used if
the display is in portrait orientation.
This causes a failure in webrtc.RTCPeerConnection that tests
the hardware encoder is used during a rtc peer connection.
This fixes the failure by setting the display orientation to
the landscape orientation in the test.

BUG=b:168059462
TEST=video.RTCPeerConnection.vp8_enc_cam on scarlet

Change-Id: I09a3af2c755b32310a33104bbe9ecf47d56c2915