tast-tests: Enable Test API Connection to work in Guest Mode
CL:2646390 introduces necessary Chromium changes to enable the Tast test
API extension to be loaded as a component extension which works in Guest
Mode. Ensure the Tast API extension is loaded into a temp directory even
when the chrome.GuestLogin option is passed in.

Additionally it adds a smoke test ui.GuestModeExtension which simply
loads Chrome in Guest mode and tries to connect to the Test API.

NOTE: This must not be submitted until CL:2646390 is uprevved into
ChromeOS to avoid the switch (--load-guest-mode-test-extension) not
being present.

BUG=chromium:1136074, chromium:1122877, chromium:1079235
TEST=tast run <DUT-IP> apps.LaunchSystemWebAppsFromURL*

Change-Id: I7dfbd24d4dc96f19715075a1570b93057f3c6798