tast-tests: Enable Test API Connection to work in Guest Mode
CL:2646390 introduces necessary Chromium changes to enable the Tast test
API extension to be loaded as a component extension which works in Guest
Mode. Ensure the Tast API extension is loaded into a temp directory even
when the chrome.GuestLogin option is passed in.

This updates the apps.LaunchSystemWebAppsFromURL test to also test from
guest mode to verify the functionality.

BUG=chromium:1136074
TEST=tast run <DUT-IP> apps.LaunchSystemWebAppsFromURL*

Change-Id: I7dfbd24d4dc96f19715075a1570b93057f3c6798