Fix test failures of arc.ChromeVirtualKeyboard.unstable.
It fails only if it ran immediately after arc.ChromeVirtualKeyboard.
The focusChange test implicitly assumes that the input focus just after
launch is not set to any view. However, it is not true in some cases.
This CL removes the assumption by explicitly setting the focus at the
beginning of the tests by reordering checks.

BUG=b:177467902
TEST=tast run ${DUT} arc.ChromeVirtualKeyboard arc.ChromeVirtualKeyboard.unstable

Change-Id: I36a9bec4c801ef587bb4bc2a44a37a6f74267af0