health.DiagnosticsRun: add NVMe hardware dep for NVMe routines
This change adds a NVMe hardware dependency to the nvme_wear_level and
nvme_self_test routines. Previously only a software dependency was used,
but many boards have some models that have NVMe storage and others that
do not. The USE flag is not enough to say for sure that the device has
NVMe storage.

BUG=b:178374608
TEST=tast run sarien-dut health.DiagnosticsRun.nvme_self_test (pass)

Cq-Depend: chromium:2647527
Change-Id: Ia2e7300897b2e25a9c8230c539900fc6daa12983
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2647528
Tested-by: Trent Begin <tbegin@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>