health.ProbeNetworkInfo: remove VPN from network device count
Chrome previously reported a VPN as a network device, even if the VPN
was not configured. After fixing https://crbug.com/1156816, this is no
longer the case. This change removes the check to ensure that the VPN is
present in the network health devices.

BUG=b:178274447
TEST=tast run sarien-dut health.ProbeNetworkInfo (pass)
This DUT was configured with only one network device.

Change-Id: I55f3469ae178403c15a19621223355b79fb42b3b