Change ui.DragMaximizedWindowPerf background window states
I've noticed that the DragMaximize / Unmaximize performance is
significantly affected by the background window size. On kevin,
current scenario (4 smaller windows aligned as tiled) achieves
~100% of smoothness, while it drops ~50% smoothness when they're
maximized.

As I checked UMA, the numbers in the real environment is closer
to the maximized pattern, and also I think this is more realistic
environment -- most users will keep windows maximized or covers
majority of the desktop, I believe.

Note: this will cause a seemingly performance regression on
this test, but I think we should accept the drop as long as
it reflects the reality.

BUG=chromium:1170544
TEST=tast run ui.DragMaximizedWindowPerf on kevin

Change-Id: I9634a15c4190e0c6938cad6cd76096314ab2d29c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2648929
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>