security.HardeningStatus: List interesting processes.
We're interested in processes that are running with privileges (like
root privileges) and can also be affected by other procesess (when
these other processes perform mounts.) Identify and list these
processes.

A follow-up CL will start dumping this information in JSON form and
reify the checks so that new checks can be added without having to
modify the main loop of the test.

BUG=chromium:1165788
TEST=Passes on betty-pi-arc and prints the correct processes.

Change-Id: If943d305b028ad10cbb41427db570d66b37fa5c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2650265
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>