tast-tests: update arc.DisableArc
The test now uses the new ui library to help with stability. It also
waits for the play store to close to better track errors related to it
not closing. Also, it properly checks the final error now.
On top of that, it has a work around to keep closing Play Store.
Investigation into why ARC isn't closing Play Store needs to be done.

BUG=b:178232263
TEST=./repeat_test.sh -r 10 -s <eve> arc.DisableArc

Change-Id: Icafbdb477963b2011f75c091528e7613e1037fa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2650487
Tested-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>