scanner: Restart printing processes at start of test
CUPS is supposed to autoconfigure newly-added IPPUSB devices.  On some
boards, this isn't happening reliably, possibly due to a previous test
leaving upstart-socket-bridge and/or ippusb_manager in a wedged state.
Restart all the printing-related processes at the beginning of the test
to ensure that we start in a known state.

BUG=b:162003158
TEST=20+ iterations of tast run scanner.* on betty and zork

Change-Id: I8962640ffc2c1ac8894ae8df28a635abe2bf1bd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2653267
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>