wifi.SimpleConnect: Promote remaining 8021x testcases.
There is a known flake in these tests (b/170368080). However, the
problem is also available in Autotest version, so it should not block
the migration process.

Promote the tests to stable so that we can stop the Autotest version
later to save some lab resources.

BUG=b:178567801
TEST=`tast list $DUT '("group:wificell" && wificell_func &&
!wificell_unstable)'` does include the tests.

Change-Id: Ib032077ba8d35a433cc39aff9c45ee3634fd1ce8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2654794
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>