tast: Cover more disconnect senarious.
The tests wifi.DisconectReason.* verify the disconnect reason reported
by wpa-supplicant. The reported disconnct reason depends on things such
as the type of the WiFi driver (full-MAC vs soft-MAC), and the senario
of how the disconnection happens. For example, in the sub-test ap_gone,
the DUT could experince a deauthentication due to STA leaving or
inactivity (BEACON LOSS).

Expnaded the disconnection reasons to cover all the possible senarios.

BUG=b:174182853 , b:176254301
TEST=run the tests wifi.DisconnectReason.* on DUTs with a full-MAC
driver.

Change-Id: I1b91a1a7d3a6b3e9ec1953963df139532707419c