[gpucuj] Add custom trace analysis code.
This adds three custom metrics based on traces: number of frames, fps,
and percentage dropped frames.

N.B. Traces are still taken from inside lacros and chromeos-chrome,
depending on which browser is being tested. This is because it seems
that dropped frames inside lacros appear as DidNotProduceFrame events
from exo, which breaks the dropped frames calculation.

BUG=chromium:1140906
TEST=tast run <dut> lacros.GpuCUJ.*

Change-Id: I34589f6f44d8df5b544f335bc206ee963826e142