crostini: Rationalize log collection
Currently the way we collect logs for failing tests is a bit of a
mess. We collect different types of logs depending on when the test
failed, give up too easily on some logs, and generally make a mess of
things.

In particular, we don't collect VM logs if the precondition fails, and
we only collect LXD logs if the precondition fails.

Instead, do this:
 - Collect VM logs always, as these are just files on disk on the host
 - Collect LXD logs if we can connect to a running VM
 - Collect container logs if we are connected to a running container
 - Always do this, whether the test fails in the precondition or later

BUG=chromium:2658782
TEST=crostini.Basic

Disallow-Recycled-Builds: test-failures
Change-Id: I24999a1f9eb9b336bfa4ecda87b4203669b7a693
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2658783
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>