video: Fix Media DevTools interactions
crrev.com/c/2648444 introduced reading video player properties
via DevTools protocol. This new mode fails to retrieve those
properties on certain bots and for the Play.*guest test cases.

A bit of investigation showed that the properties are indeed
present, but for some reason they are not correctly retrieved.
This could mean a bug in the underlying github.com/mafredri/cdp
library or elsewhere, I don't know.

This CL works around the problem by giving it another chance:
when both kIsPlatformVideoDecoder and kVideoDecoderName player
properties are not filled in at all, this CL allows for giving
it another try. Local experiments on reks shows that this does
the trick nicely, i.e. the properties are read the second time
around.

BUG=b:177104508
TEST=tast run reks video.Play.*guest

Change-Id: I58b6f7e1f190fb8a0dc067f6a37723e5f6a3f53d