Add more logging to CB->phone test
Add some logging for each major sharing step to make it easier to tell
what's going on in the logs. Also change the deferred Android cleanup to
prevent an unnecessary call to CancelReceivingFile, which pollutes the
logs with a Java stack trace if we call it after sharing succeeds.

BUG=chromium:1172676
TEST=tast run <dut> neabyshare.CrosSenderAndroidReceiver.small_jpg

Change-Id: Ic69fd485d8f4a5ccff2e46951786ae5e6ba3a484
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2664853
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>