crostini: Fix panic on chrome failure
Currently the logs collection we do when we fail to set up the
precondition assumes we connected to Chrome, which may not be the
case. If we didn't we can't meaningfully collect any VM logs (and none
would have been produced anyway), so just abort early in this case.

BUG=none
TEST=none

Disallow-Recycled-Builds: test-failures
Change-Id: I2b6feac607e9790cb819bd3babdca6cddab0ccc4