wificell: include option for advertising multiple BSSIDs
Each phy can advertise multiple BSSIDs on the same channel, and hostapd
has an option for this that we haven't made use of until now. Some tests
will need this going forward, so add it here. You can check how many
SSIDs can be advertised by looking at the 'valid interface combinations'
section of `iw list`.

Note that there are several other options that can be parameterized
along with the SSID, and we can probably also run DHCP on the additional
interfaces as well, but let's add them as dummy networks for now since
there's no immediate need for those features.

BUG=b:177478042
TEST=ran with this option in a few roam tests, checked the scan results
to verify that the DUT was indeed picking up the other SSIDs, ifconfig
on the AP side showed the additional interface.

Change-Id: I1ef0faacc9a718e535ba5961c4599d43827aee90