tast-tests: update scanner.ScanESCLIPP golden image
The scanner.ScanESCLIPP tast tests compare an image returned by the
virtual-usb-printer to a golden image in the cloud. The
virtual-usb-printer used to return a canned image, but now it generates
images from an algorithm with the scan settings as inputs. Update the
golden image in the cloud to reflect the new expected image.

BUG=b:178419192
TEST=tast run --build $DUT_IP scanner.ScanESCLIPP*

Change-Id: I2bd4b6ced7b8ffc33e935081210ec0320a18e30f