tast-tests: pkg.go.dev file naming
It seems that pkg.go.dev does not like files with '*' in the name. This
may be the reason it can't show docs for the tast-tests repo. Besides
these file changes, also copied the file to the new location in gcs.

BUG=chromium:1173778
TEST=tast run <eve> apps.LaunchGallery*

Change-Id: I8a559d6c6e63a4b8e65d07b76a0b92eb795045da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2669989
Tested-by: Brendan Hansknecht <bhansknecht@chromium.org>
Auto-Submit: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Brendan Hansknecht <bhansknecht@chromium.org>