Tast-tests: Fix failure accessing TPM when DUT is not connected.
When running tast.example.ReconnectToDUT, the connection to the DUT
is closed resulting in post test fix up to fail when trying to access
the TPM status.
Post test fix up should not assume that there is still an active
connection to the DUT upon test exit. Post test fixup need to
re-connect to the DUT if a connection does not exist.

BUG=chromium:1167941

TEST=tast run localhost:9222 example.ReconnectToDUT

Change-Id: I41c5838e28dcbe7318cc751c114d43592d42eb58