tast-tests: Fixes SELinux context mismatch errors
Test - security.SELinuxFilesDataDir has been failing in
few instances as some files or dirs are not found.
This occurs mainly for temporary files - which are
not found during the race between find and matchpathcon
command.

This CL accounts for such not found files and dirs so
that they don't result in incorrect SELinux context
mismatch errors.

Example failure -
https://b.corp.google.com/issues/162202740#comment2

Tastboard Results -
http://tastboard/test/security.SELinuxFilesDataDir

BUG=b:150782430
TEST=tast run $IP security.SELinuxFilesDataDir

Change-Id: I1ff2f9c646485e4203525df44973359f30c403a7