power: Log battery info and report discharge to Crosbolt.
Start gathering the data that would be useful to diagnose the random DUT
shutdowns that occasionally happen.

When starting and stopping power measurements log the battery capacity
in % and battery energy in Wh. Additionally log battery discharge in
both % and Wh, and report the discharge in mWh to Crosbolt.

BUG=b:176514670
TEST=tast run -buildbundle=crosint DUT \
      platform.MLBenchmark.smartdim_no_nnapi

Cq-Depend: chrome-internal:3573027
Change-Id: I92bef043f3606648034975d880946faacda7d087