Lacros: Fix permission issue of test extension dir
Lacros chrome runs as the chronos user, and the outer most test
extension directory was created in a way that chronos user doesn't have
read access to it, thus Lacros tast tests fail to load the test
extension due to fail to read the manifestion file.

This CL fixes the issue by granting proper access to the directory.

BUG=chromium:1173889
TEST=Ran Lacros.ShelfLaunch and Lacors.AppLauncherLaunch on DUT

Change-Id: Iff8e362ba8a9b1873157b24dd10c23858fe9baee