Store chrome log into a file per test
When a test is failing or runs into a weird situation, we usually
check with the chrome log file to see if there's something strange.
However, it is quite cumbersome to find out which log file to
be seen from the test log, especially in CI because there are a
lot of tests are running, and they share the preconditions or
failures.

To make the debugging easier, I think something like this would be
helpful -- storing the chrome log during a test into a file
"chrome.log" in per-test output directory.

BUG=chromium:1174574
TEST=tast run <DUT> ui.DesksAnimationPerf ui.OverviewScrollPerf

Change-Id: I2d2082ff848bbfcf58845e0e9fee7015020328a6