arc.MemoryChromeOSPerf mmap smaller ranges
There are failures in the mmap, which might be because we are running
out of VA, and things are fragmented enough that 50MB ranges don't have
any place to fit.

This CL makes all mmap ranges 1MB or less and also logs how much we have
allocated, and how much VA the tast process used if we do have a
failure.

BUG=b:166750882
TEST=tast run dut arc.MemoryChromeOSPerf.vm

Change-Id: I42dc8ad6516c332bb04f2a0fecdb27c1a4dcab33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2672631
Tested-by: Charles Dueck <cwd@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Charles Dueck <cwd@chromium.org>