Be more exact when scanning for Chrome crash files
Previously, ChromeCrash tests just looked for a file that ended with the
correct PID. However, that's vulnerable to having previous crashes (such
a selinux violation) that used that PID.

Be more picky about the filename to avoid this issue.

Note: In theory, we could also pick up a DumpWithoutCrashing crash and
incorrectly return that. However, in that case the tests will still
pass because there will still be a .dmp file. We can tolerate an
occasional flaky-pass but not a flaky-fail.

BUG=chromium:1146672
Fixed: chromium:1146672
TEST=ui.ChromeCrashNotLoggedIn.gpu_process_crashpad_mock_consent passed

Change-Id: I9245da676a2d792b05e5349debd3df35dab52753
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2676609
Reviewed-by: Michael Irani <michaelirani@google.com>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Auto-Submit: Ian Barkley-Yeung <iby@chromium.org>