Disable systrace for UI performance tests
It seems that kernel gets panicked on dedede devices when some
of the UI operations are conducted with chrome-tracing enabled
with Systrace flag=true.

In my understanding, systrace information is not quite important
to those test cases, and so I think it's okay to uniformly
disable it for the time being.

BUG=chromium:1162385
TEST=tast run <DUT> '("crosbolt_perbuild" && "name:ui.*" && !"name:*CUJ*")'

Change-Id: I078f26ef5d339666544afb492f9c77e6820a212a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2677158
Commit-Queue: Jun Mukai <mukai@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>