Fix ui.WindowControl to support the new behavior.
Based on crbug.com/1158548, when the maximized window is dragged
back to the top position, it needs to wait a bit (more than
800msecs?) before releasing the mouse. Othrewise, the window
stays in normal state, not gets back to maximized.

Those 2 tests (ui.WindowControl and ui.DragMaximizedWindowPerf)
conducts those operations in Tast. They need the waiting mechanism
to support this new behavior.

BUG=chromium:1174874
TEST=tast run <DUT> ui.WindowControl ui.DragMaximizedWindowPerf

Change-Id: I8a28135f44f59b882a1339af707463f82070d673