(Nearby Share) Fix TCP read deadline for slow-to-respond RPCs
The read deadline for the TCP conn to the Nearby Snippet is currently 10
seconds for all RPC responses. However some RPCs (eventWaitAndGet,
settings getters) will wait up until a specified timeout to respond.
This is causing the read to occasionally time out when we are waiting
for eventWaitAndGet's response, particularly when we are waiting for the
share to finish. To fix this, clientRPCResponse now accepts a timeout
parameter for how long to wait for a response from the snippet.

BUG=chromium:1173847
TEST=tast run <lab dut> nearbyshare.*

Change-Id: Ic8031a3a3111f51726b87edc1455ec701da45adc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2680029
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>