ShillCellularEnableAndConnect: Improve reliability
This adds a wait for the 'Scanning' property to be false after
an Enable, and increases the Sleep time after Disable.
(The Sleep time still has a TODO requiring a Shill fix, but
the current time of 200ms was insufficient for coral devices).

This also includes some cleanup, adding a generic
'WaitForShillProperty' function to property_holder.go.

BUG=b:177103167
TEST=tast run ${DUT} network.ShillCellularEnableAndConnect

Change-Id: Ie01c232296544d7e068c04e072850c447ecbcec0