wifi: Promote ChannelScanDwellTime to stable
As discussed on b/175541088 the test appears to be performing
at least as well as network_WiFi_ChannelScanDwellTime.

The bulk of test failures appear to be caused by failures of
the pcap router to record specific packets of interest, but this
is a thing we knew could happen occasionally and the rate at which
it is occurring seems broadly reasonable.

BUG=b:175541088
TEST=Stainless shows reasonable pass rate over last 14 days

Change-Id: I95a8734b4c0ecb93b4491fa537c9877e867f1bef