tast-tests: syzkaller: rewrite data/enabled_syscalls.txt
Seperate descriptions corresponding to drivers based on board
architecture, so that descriptions for i915 only run against amd64 and
descriptions for mali run only against arm64.

Move enabled_syscalls.txt to enabled_syscalls.json to allow for tracking
the following information on a per-driver basis; architecture to run on,
startup commands to run, and syscalls corresponding to the
driver/subsystem.

BUG=b:176826954
TEST=tast -verbose run localhost:2200 syzkaller.Wrapper

Change-Id: I5c07c129cb55f8a92edc492c0f85a2776069fc72