tast-tests: fix unexpected error of hardware.SensorActivity
Originally, the hardware.SensorActivity will tests the on-body detection
if the activity sensor presents, but if the DUT is supporting other
activities but on-body detection, the test will fail. Thus, I modify the
test so that if the on-body detection not present on the DUT, it will
skip the on-body detection test.

BUG=b:177660658
TEST=tast run <ip> hardware.SensorActivity
     (with and without on-body detection presents)

Change-Id: I5bbbb08feb2162a422319db6e803bce1a567e948
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2678837
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Ching-Kang Yen <chingkang@chromium.org>
Commit-Queue: Ching-Kang Yen <chingkang@chromium.org>