tast: Add send_chan_switch disconnect scenario.
The sub-test send_chan_switch triggers the AP to send a CSA frame to the
DUT and verifies the disconnect reason. In this case, the DUT switches
to the new channel but it does not find the AP, becasue the AP does not
switch channel, it only sends the CSA frame.

Boards with Marvell WiFi chips have full-MAC drivers and discionnect with
one of the following reasons:
1- wpasupplicant.DisconnReasonLGDisassociatedInactivity.
2- wpasupplicant.DisconnReasonLGDeauthSTALeaving (Added in this CL).
3- wpasupplicant.DisconnReasonUnknown.

BUG=b:174182853 , b:176254301
TEST=run the test wifi.DisconnectReason.send_chan_switch on DUTs with Marvell WiFi
chip.

Change-Id: I589700514efaf1ece25e4f49c63a942da5bda55a