Lacros: Increase Lacros launch timeout
Lacros tast tests have been consistently failing on low-end
e2-standard-8 VMs because the tests didn't see a Lacros window being
visible after waiting for 10s. A hypothesis is that on slow VMs, Lacros
takes a significant longer time to launch, so this CL increases launch
timeout to 1m and see if it fixes the problem.

BUG=chromium:1168664
TEST=Ran Lacros tast tests

Change-Id: Ia36d74925147411ba829f5b2b86221f9358b1cef