video/PlatformEncoding: add support for FPS estimation
This CL adds the code to calculate the produced encoded bitrate
and add a Perf entry with the deviation from the target one.
The new new entry is a deviation percentage, e.g. -5%, +3%.

(There's no CPFE requirement as to how much would this need to
be but we should soon add one. FTR WebRTC sets [-95%,115%] for
VP8).

BUG=b:178542866
TEST=tast run video.PlatformEncoding.* on volteer and kukui

Change-Id: I08b61606e0d750404fa592354bb95152e94069c0