video/PlatformEncoding: add support for produced bitrate estimation
This CL adds the code to calculate the produced encoded bitrate
and adds a Perf entry with the deviation from the target one.
The new new entry is a deviation percentage, e.g. -5%, +3%.

(There's no CPFE requirement as to how much would this need to
be but we should soon add one. FTR WebRTC sets [-100%,+15%] for
VP8).

BUG=b:178542866
TEST=tast run video.PlatformEncoding.* on volteer and kukui

Change-Id: I08b61606e0d750404fa592354bb95152e94069c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2682981
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>