tast-tests: Wait for Guest process to start to identify correct PID
The current way to identify if a Chrome renderer is available does not
work for Guest processes. The method is to wait for the cryptohome user
mount then to get the PID but it seems this is not sufficient when
logging into a Guest session.

This introduces a new method which waits for the Chrome renderer process
with the command line argument --login-user=$guest to appear and will
poll indefinitely at 10ms intervals until it starts up.

BUG=chromium:1174939
TEST=tast run <DUT-IP> ui.GuestModeExtension

Change-Id: I22a35a2d553e4448d2572fe8794cbb4813f50247